pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lawgraph"
version = "0.1.0"
description = "Pipelines voor NL/EU wetgeving en rechtspraak naar een ArangoDB knowledge graph."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }

authors = [
  { name = "You", email = "you@example.com" }
]

dependencies = [
  "pyyaml>=6.0.0",
  "requests>=2.31.0",
  "python-arango>=7.6.0",
  "python-dotenv>=1.0.0",
  "typing-extensions>=4.12.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]

[project.urls]
Homepage = "https://example.com/lawgraph"
Source = "https://example.com/lawgraph.git"

[project.scripts]
lawgraph-parliament = "lawgraph.cli.parliament:main"
lawgraph-case-law   = "lawgraph.cli.case_law:main"
lawgraph-eu-law     = "lawgraph.cli.eu_law:main"
lawgraph-retrieve-tk = "lawgraph.cli.retrieve_tk:main"
lawgraph-retrieve-rechtspraak = "lawgraph.cli.retrieve_rechtspraak:main"
lawgraph-retrieve-eurlex = "lawgraph.cli.retrieve_eurlex:main"
lawgraph-retrieve-bwb = "lawgraph.cli.retrieve_bwb:main"
lawgraph-retrieve-all = "lawgraph.cli.retrieve_all:main"
lawgraph-normalize-tk = "lawgraph.cli.normalize_tk:main"
lawgraph-normalize-rechtspraak = "lawgraph.cli.normalize_rechtspraak:main"
lawgraph-normalize-eurlex = "lawgraph.cli.normalize_eurlex:main"
lawgraph-normalize-bwb = "lawgraph.cli.normalize_bwb:main"
lawgraph-normalize-all = "lawgraph.cli.normalize_all:main"
lawgraph-strafrecht-seed = "lawgraph.cli.strafrecht_seed:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
